Changes from Harbour Project https://github.com/harbour/core/commit/d54bb0e3f0cdd5f75856012c8ae75b63e73105c8

+++ src/rtl/replic.cpp
@@ -61,7 +61,7 @@ HB_FUNC( REPLICATE )
       if( nLen > 0 && nTimes > 0 ) {
          if( nTimes == 1 ) {
             hb_itemReturn(pItem);
-         } else if( static_cast<double>(nLen) * nTimes < HB_SIZE_MAX ) {
+         } else if( static_cast<double>(nLen) * nTimes < static_cast<double>(HB_SIZE_MAX) ) {
             auto szText = hb_itemGetCPtr(pItem);
             HB_SIZE nSize = nLen * nTimes;
             char * szResult, * szPtr;
